docs: update README to emphasize experimental status and usage warnings (#6)
* docs: update README to emphasize experimental status and usage warnings

* style: satisfy clang-format and clang-tidy CI checks

- apply clang-format across src/ and test/ (Format Check)
- fix clang-tidy findings (Tidy Check):
  - braces-around-statements in geom_model bbox loop
  - explicit widening cast for triangle_count * 3 (payload)
  - explicit double cast in centroid division (geom_analysis)
  - hoist if-init declaration; add <algorithm> for std::minmax
- set CMAKE_CXX_STANDARD 17 so the tidy pass parses as C++17


* ci: fix extension name, relicense to TU Delft, add justfile

- MainDistributionPipeline: extension_name quack -> three_d (was failing the build)
- remove dead ExtensionTemplate.yml (template-only, jobs always skipped)
- LICENSE: MIT copyright -> Delft University of Technology
- add justfile with build/test/format/tidy/ci shortcuts mirroring CI


* chore: upgrade DuckDB to v1.5.4

- MainDistributionPipeline: duckdb_version / ci_tools_version / reusable-workflow refs v1.5.3 -> v1.5.4
- bump duckdb submodule to v1.5.4 (08e34c4)
- bump extension-ci-tools submodule to v1.5.4 (b777c70)


---------

+++ src/kernel/payload.cpp
@@ -105,8 +105,8 @@ void ValidatePayloadModel(const SolidModel &model, uint32_t vertex_count, uint32
 		}
 	}
 
-	if (model.vertices.size() != vertex_count || model.SolidCount() != solid_count || model.ShellCount() != shell_count ||
-	    model.FaceCount() != face_count || model.RingCount() != ring_count ||
+	if (model.vertices.size() != vertex_count || model.SolidCount() != solid_count ||
+	    model.ShellCount() != shell_count || model.FaceCount() != face_count || model.RingCount() != ring_count ||
 	    model.TriangleCount() != triangle_count) {
 		throw std::runtime_error("SOLID_3D payload: header counts do not match payload body");
 	}
@@ -190,7 +190,7 @@ std::vector<uint8_t> SerializePayload(const SolidModel &model) {
 	writer.WriteArray(model.ring_vertex_indices.data(), model.ring_vertex_indices.size());
 
 	// Data arrays: triangle vertex indices
-	writer.WriteArray(model.triangle_vertex_indices.data(), triangle_count * 3);
+	writer.WriteArray(model.triangle_vertex_indices.data(), static_cast<size_t>(triangle_count) * 3);
 
 	// Validation cache
 	writer.Write<uint32_t>(model.validation.open_edge_count);
